readRAST.Rd exmpl: take control of mapsets; only write to a custom one *
In the course of this fiddling with mapsets, it was seen (by accident) that the mapset
region settings are not respected by the GDAL driver.

This only affects raster.

This is a no-go for the GDAL-GRASS raster driver as long as it doesn't expose an option to
respect the mapset's current region, or at least to set a region.

A feature request will be made at the driver's repo, but as long as it's not implemented,
we cannot consider using the driver for rasters since results with and without using the
driver must of course have the same extent and resolution.

For sake of completeness, this commit takes control of the region too, in order to
guarantee that things always work. But this will be dropped from the examples in a next
commit, followed by dropping driver-related code from read_RAST() and write_RAST().