readRAST.Rd exmpl: don't control region; should work with any region *
It was seen (by accident) that the mapset region settings are not respected by the GDAL
driver.

This only affects raster.

This is a no-go for the GDAL-GRASS raster driver as long as it doesn't expose an option to
respect the mapset's current region, or at least to set a region.

A feature request will be made at the driver's repo, but as long as it's not implemented,
we cannot consider using the driver for rasters since results with and without using the
driver must of course have the same extent and resolution.

So next is dropping driver-related code from read_RAST() and write_RAST().