read_RAST(), write_RAST(): drop GDAL-driver use (revert e1f026d, a06e782, 4389249) *
In the course of this fiddling with mapsets, it was seen (by accident) that the mapset
region settings are not respected by the GDAL driver.

This only affects raster.

This is a no-go for the GDAL-GRASS raster driver as long as it doesn't expose an option to
respect the mapset's current region, or at least to set a region.

A feature request has been made at the driver's repo
(https://github.com/OSGeo/gdal-grass/issues/49). As long as it's not implemented, we
cannot consider using the driver for rasters since results with and without using the
driver must of course have the same extent and resolution.